xenomai-2.4.6

+++ include/rtdm/rtdm.h
@@ -49,12 +49,13 @@
 
 #ifdef __KERNEL__
 
+#include <linux/types.h>
 #include <linux/fcntl.h>
 #include <linux/ioctl.h>
 #include <linux/sched.h>
 #include <linux/socket.h>
 
-typedef size_t socklen_t;
+typedef u32 socklen_t;
 typedef struct task_struct rtdm_user_info_t;
 
 #else /* !__KERNEL__ */
